requirements.txt
langchain==0.1.20
langchain-openai==0.1.7
langgraph==0.0.41
python-dotenv>=1.0.1
tqdm>=4.66.1
openai>=1.30.1